Refactored regulaer expressions
All the regular expressions used in the extraction operation were combined into a single data structure - since they were so many with a significant amount of repetition. This is to ease any future maintenance as well as for further abstraction. In doing so, some of the regexes were fine-tuned and associated code-breakage fixed.